+++ CHANGES.md
@@ -2,7 +2,11 @@
 
 ## 0.2.7 (alpha)
 - add `Geometry.indexCount()` to control the number of indices drawn in an indexed draw
-- add `GText.characters(int)` to control the number of characters rendered on a GText
+- New `GText` features
+  - `GText.characters(int)` to control the number of characters rendered on a GText
+  - `GText.width(int)` to set text wrap width
+  - `GText.align(int)` to set left/center/right alignment
+  - `GText.size(int)` to scale text size while respecting wrap and alignment
 - Bugfix: incorrect transparency sorting on scenes with a large number of materials
 
 ## 0.2.6 (alpha)